data_service/requirements.txt
Flask
PyMySQL
mysql-connector-python
python-dotenv
jwt
PyJWT